+++ src/KeyVault/KeyVault/Models/Key/PSKeyVaultKey.cs
@@ -68,6 +68,9 @@ namespace Microsoft.Azure.Commands.KeyVault.Models
 
             // Key additional properties
             Attributes = new PSKeyVaultKeyAttributes(key);
+
+            // External Key Manager (EKM) backed key id, if any.
+            ExternalKeyId = key.Properties.ExternalKey?.Id;
         }
         #endregion
 
@@ -95,6 +98,9 @@ namespace Microsoft.Azure.Commands.KeyVault.Models
                 return Attributes.ReleasePolicy;
             }
         }
+
+        // External Key Manager (EKM) key id for an external key on a Managed HSM. (Preview)
+        public string ExternalKeyId { get; set; }
 
         #endregion
     }